+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.116] - Not released
+## [1.116] - 2023-03-10
 ### Added
 - Rich event preview on Notifications page. Every event now has a user picture
   and text of comment or post (if this event is comment- or post-related).